feat: Add support for linking failure reports
Introduces functionality to link external documents to traction failure records.
A new `documentlink` field is added to the `TractionFailure` type.
The `FailuresTable` component is updated to display a clickable link icon next to the locomotive number if a `documentlink` is present.
The `FailureDetailsModal` component is also updated to show the link.
Documentation in README.md is updated to reflect this new feature.